LRP-4526 Cast to int to keep the same logic as before

+++ tools/release/ext/ext-impl/src/com/liferay/portal/license/LicenseManager.java
@@ -942,7 +942,7 @@ public class LicenseManager {
 				pos = 0;
 			}
 			else {
-				pos += c;
+				pos += (int)c;
 			}
 		}
 	}
